Move isDonorEligibleToDonate to DonorConstraintChecker
