+++ src/service/DonorConstraintChecker.java
@@ -1,14 +1,14 @@
 package service;
 
 import javax.persistence.NoResultException;
-
+import model.donation.Donation;
 import model.donor.Donor;
-
+import model.packtype.PackType;
+import org.joda.time.DateTime;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
 import repository.DonationRepository;
 import repository.DonorDeferralRepository;
 import repository.DonorRepository;
@@ -23,6 +23,8 @@ public class DonorConstraintChecker {
     private DonationRepository donationRepository;
     @Autowired
     private DonorDeferralRepository donorDeferralRepository;
+    @Autowired
+    private DonorDeferralStatusCalculator donorDeferralStatusCalculator;
     
     public boolean canDeleteDonor(long donorId) throws NoResultException {
         Donor donor = donorRepository.findDonorById(donorId);
@@ -39,6 +41,40 @@ public class DonorConstraintChecker {
             return false;
         }
 
+        return true;
+    }
+    
+    public boolean isDonorEligibleToDonate(long donorId) {
+        
+        Donor donor = donorRepository.findDonorById(donorId);
+        
+        if (donor.getDonations() != null) {
+
+            for (Donation donation : donor.getDonations()) {
+    
+                PackType packType = donation.getPackType();
+    
+                if (!packType.getCountAsDonation()) {
+                    // Don't check period between donations if it doesn't count as a donation
+                    continue;
+                }
+    
+                // Work out the next allowed donation date
+                DateTime nextDonationDate = new DateTime(donation.getDonationDate())
+                        .plusDays(packType.getPeriodBetweenDonations())
+                        .withTimeAtStartOfDay();
+                
+                // Check if the next allowed donation date is after today
+                if (nextDonationDate.isAfter(new DateTime().withTimeAtStartOfDay())) {
+                    return false;
+                }
+            }
+        }
+
+        if (donorDeferralStatusCalculator.isDonorCurrentlyDeferred(donor)) {
+            return false;
+        }
+        
         return true;
     }
