+++ test/service/DonationConstraintCheckerTests.java
@@ -1,15 +1,10 @@
 package service;
 
 import static helpers.builders.DonationBuilder.aDonation;
-import static helpers.builders.DonorBuilder.aDonor;
-import static helpers.builders.PackTypeBuilder.aPackType;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.mockito.Mockito.when;
-import java.util.Date;
 import model.donation.Donation;
-import model.donor.Donor;
-import org.joda.time.DateTime;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
@@ -18,13 +13,11 @@ import org.mockito.runners.MockitoJUnitRunner;
 import repository.BloodTestResultRepository;
 import repository.ComponentRepository;
 import repository.DonationRepository;
-import repository.DonorRepository;
 
 @RunWith(MockitoJUnitRunner.class)
 public class DonationConstraintCheckerTests {
     
     private static final long IRRELEVANT_DONATION_ID = 17;
-    private static final long IRRELEVANT_DONOR_ID = 96;
 
     @InjectMocks
     private DonationConstraintChecker donationConstraintChecker;
@@ -34,10 +27,6 @@ public class DonationConstraintCheckerTests {
     private BloodTestResultRepository bloodTestResultRepository;
     @Mock
     private ComponentRepository componentRepository;
-    @Mock
-    private DonorRepository donorRepository;
-    @Mock
-    private DonorDeferralStatusCalculator donorDeferralStatusCalculator;
     
     @Test
     public void testCanDeleteDonationWithDonationWithNotes_shouldReturnFalse() {
@@ -118,98 +107,6 @@ public class DonationConstraintCheckerTests {
         boolean canDelete = donationConstraintChecker.canUpdateDonationFields(IRRELEVANT_DONATION_ID);
         
         assertThat(canDelete, is(true));
-    }
-    
-    @Test
-    public void testIsDonorEligibleToDonateWithNonDeferredDonorWithoutDonations_shouldReturnTrue() {
-        
-        Donor donor = aDonor().withId(IRRELEVANT_DONOR_ID).build();
-        
-        when(donorRepository.findDonorById(IRRELEVANT_DONOR_ID)).thenReturn(donor);
-        when(donorDeferralStatusCalculator.isDonorCurrentlyDeferred(donor)).thenReturn(false);
-        
-        boolean isEligible = donationConstraintChecker.isDonorEligibleToDonate(IRRELEVANT_DONOR_ID);
-        
-        assertThat(isEligible, is(true));
-    }
-    
-    @Test
-    public void testIsDonorEligibleToDonateWithDeferredDonorWithoutDonations_shouldReturnFalse() {
-        
-        Donor donor = aDonor().withId(IRRELEVANT_DONOR_ID).build();
-        
-        when(donorRepository.findDonorById(IRRELEVANT_DONOR_ID)).thenReturn(donor);
-        when(donorDeferralStatusCalculator.isDonorCurrentlyDeferred(donor)).thenReturn(true);
-        
-        boolean isEligible = donationConstraintChecker.isDonorEligibleToDonate(IRRELEVANT_DONOR_ID);
-        
-        assertThat(isEligible, is(false));
-    }
-    
-    @Test
-    public void testIsDonorEligibleToDonateWithNonDeferredDonorWithDonationNotCountedAsDonation_shouldReturnTrue() {
-        
-        Donor donor = aDonor()
-                .withId(IRRELEVANT_DONOR_ID)
-                .withDonation(aDonation()
-                        .withPackType(aPackType().withCountAsDonation(false).build())
-                        .build())
-                .build();
-        
-        when(donorRepository.findDonorById(IRRELEVANT_DONOR_ID)).thenReturn(donor);
-        when(donorDeferralStatusCalculator.isDonorCurrentlyDeferred(donor)).thenReturn(false);
-        
-        boolean isEligible = donationConstraintChecker.isDonorEligibleToDonate(IRRELEVANT_DONOR_ID);
-        
-        assertThat(isEligible, is(true));
-    }
-    
-    @Test
-    public void testIsDonorEligibleToDonateWithNonDeferredDonorWithDonationNotOverlappingPeriod_shouldReturnTrue() {
-        
-        Date previousDonationDate = new DateTime().minusDays(3).toDate();
-        
-        Donor donor = aDonor()
-                .withId(IRRELEVANT_DONOR_ID)
-                .withDonation(aDonation()
-                        .withDonationDate(previousDonationDate)
-                        .withPackType(aPackType()
-                                .withCountAsDonation(true)
-                                .withPeriodBetweenDonations(3)
-                                .build())
-                        .build())
-                .build();
-        
-        when(donorRepository.findDonorById(IRRELEVANT_DONOR_ID)).thenReturn(donor);
-        when(donorDeferralStatusCalculator.isDonorCurrentlyDeferred(donor)).thenReturn(false);
-        
-        boolean isEligible = donationConstraintChecker.isDonorEligibleToDonate(IRRELEVANT_DONOR_ID);
-        
-        assertThat(isEligible, is(true));
-    }
-    
-    @Test
-    public void testIsDonorEligibleToDonateWithNonDeferredDonorWithDonationOverlappingPeriod_shouldReturnFalse() {
-        
-        Date previousDonationDate = new DateTime().minusDays(3).toDate();
-        
-        Donor donor = aDonor()
-                .withId(IRRELEVANT_DONOR_ID)
-                .withDonation(aDonation()
-                        .withDonationDate(previousDonationDate)
-                        .withPackType(aPackType()
-                                .withCountAsDonation(true)
-                                .withPeriodBetweenDonations(5)
-                                .build())
-                        .build())
-                .build();
-        
-        when(donorRepository.findDonorById(IRRELEVANT_DONOR_ID)).thenReturn(donor);
-        when(donorDeferralStatusCalculator.isDonorCurrentlyDeferred(donor)).thenReturn(false);
-        
-        boolean isEligible = donationConstraintChecker.isDonorEligibleToDonate(IRRELEVANT_DONOR_ID);
-        
-        assertThat(isEligible, is(false));
     }
 
 }
